Refactor project and remove unused code
A significant refactor of the project was performed to streamline and remove unnecessary or unused classes, methods, or features. This included updating the logic of some existing classes to ensure efficiency and coherence. The logic of managing Torrent downloads is now handled by ITorrentService provided by the Service project, which simplified the Gui project. Referential packages and dependencies that are no longer needed have also been removed to optimize the project setup.